requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asgiref==3.3.1            # via django
colorlog==4.0.2           # via -r requirements.in
django-environ==0.4.5     # via -r requirements.in
django==3.1.3             # via -r requirements.in
pytz==2018.9              # via django
sqlparse==0.4.1           # via django
